added defining current place

+++ app/src/main/java/com/chumazkiyway/weather/models/Location.kt
@@ -1,3 +1,3 @@
 package com.chumazkiyway.weather.models
 
-data class Location(var cityName: String, var long: Float, var lat: Float)
+data class Location(var cityName: String, var long: Double, var lat: Double)
